Add metadata check and add flag to suppress deletion of downloads
In order for the checks not to fail the metadata on the C3 server has to be correct.
There were previous problems on Hycom with end of the day time being 23,
although it should be 21. This is fixed by using a cronjob to correct the metadata.

+++ ocean_navigation_simulator/environment/ArenaFactory.py
@@ -292,6 +292,7 @@ class ArenaFactory:
         points: Optional[List[SpatialPoint]] = None,
         throw_exceptions: bool = False,
         c3: Optional = None,
+        remove_files_if_corrupted: bool = True,
     ) -> List[str]:
         """
         helper function for thread-safe download of available current files from c3
@@ -360,7 +361,11 @@ class ArenaFactory:
 
         # Step 4: Download files thread-safe
         downloaded_files = ArenaFactory._download_filelist(
-            files=files, download_folder=download_folder, throw_exceptions=throw_exceptions, c3=c3
+            files=files,
+            download_folder=download_folder,
+            throw_exceptions=throw_exceptions,
+            c3=c3,
+            remove_files_if_corrupted=remove_files_if_corrupted,
         )
 
         return downloaded_files
@@ -436,7 +441,9 @@ class ArenaFactory:
         )
 
     @staticmethod
-    def _download_filelist(files, download_folder, throw_exceptions, c3):
+    def _download_filelist(
+        files, download_folder, throw_exceptions, c3, remove_files_if_corrupted=True
+    ):
         """
         thread-safe download with corruption and file size check
         Arguments:
@@ -480,34 +487,34 @@ class ArenaFactory:
                         )
                     else:
                         # check valid xarray file and meta length
-                        # try:
-                        #     grid_dict_list = get_grid_dict_from_file(
-                        #         temp_folder + filename, currents="total"
-                        #     )
+                        try:
+                            # Check valid xarray
+                            grid_dict_list = get_grid_dict_from_file(
+                                temp_folder + filename, currents="total"
+                            )
+                            # Check valid meta data
+                            if (
+                                file.subsetOptions.timeRange.start < grid_dict_list["t_range"][0]
+                                or grid_dict_list["t_range"][-1] < file.subsetOptions.timeRange.end
+                            ):
+                                error = "File shorter than declared in meta: filename={filename}, meta: [{ms},{me}], file: [{gs},{ge}]".format(
+                                    filename=filename,
+                                    ms=file.subsetOptions.timeRange.start.strftime(
+                                        "%Y-%m-%d %H-%M-%S"
+                                    ),
+                                    me=file.subsetOptions.timeRange.end.strftime(
+                                        "%Y-%m-%d %H-%M-%S"
+                                    ),
+                                    gs=grid_dict_list["t_range"][0].strftime("%Y-%m-%d %H-%M-%S"),
+                                    ge=grid_dict_list["t_range"][-1].strftime("%Y-%m-%d %H-%M-%S"),
+                                )
 
-                        #     if (
-                        #         file.subsetOptions.timeRange.start < grid_dict_list["t_range"][0]
-                        #         or grid_dict_list["t_range"][-1] < file.subsetOptions.timeRange.end
-                        #     ):
-                        #         error = "File shorter than declared in meta: filename={filename}, meta: [{ms},{me}], file: [{gs},{ge}]".format(
-                        #             filename=filename,
-                        #             ms=file.subsetOptions.timeRange.start.strftime(
-                        #                 "%Y-%m-%d %H-%M-%S"
-                        #             ),
-                        #             me=file.subsetOptions.timeRange.end.strftime(
-                        #                 "%Y-%m-%d %H-%M-%S"
-                        #             ),
-                        #             gs=grid_dict_list["t_range"][0].strftime("%Y-%m-%d %H-%M-%S"),
-                        #             ge=grid_dict_list["t_range"][-1].strftime("%Y-%m-%d %H-%M-%S"),
-                        #         )
-
-                        # except Exception:
-                        #     error = f"Corrupted file: '{filename}'."
-                        pass
+                        except Exception:
+                            error = f"Corrupted file: '{filename}'."
 
                     if error and throw_exceptions:
                         raise CorruptedOceanFileException(error)
-                    elif error:
+                    elif error and remove_files_if_corrupted:
                         os.remove(temp_folder + filename)
                         logger.warning(error)
                         continue
